requirements.txt
streamlit
numpy
Pillow
matplotlib
keras